git-svn-id: file:///Users/greg/svn2/CornerClick/trunk@575 033d73b4-3e0c-0410-a4cb-82f55ea64545

+++ clickBG/BubbleAction.h
@@ -11,16 +11,16 @@
 
 @interface BubbleAction : NSObject {
 	NSArray *actions;
-	float spacingSize;
+	CGFloat spacingSize;
 	NSSize preferredSize;
 }
 
 + (void) initialize;
 + (NSImage *) triangleImage;
-- (id) initWithSpacing:(float) space;
-- (id) initWithSpacing:(float) space
+- (id) initWithSpacing:(CGFloat) space;
+- (id) initWithSpacing:(CGFloat) space
 					 andActions:(NSArray *) theActions;
-- (void) setSpacingSize: (float) size;
+- (void) setSpacingSize: (CGFloat) size;
 - (NSSize) preferredSize;
 - (void) calcPreferredSize;
 - (void) drawInRect: (NSRect) rect;
