+++ clickBG/ClickAction.h
@@ -6,10 +6,10 @@
 @interface ClickAction : NSObject <NSCopying>
 {
     NSString* myString;
-    int theType;
-    int theCorner;
-    int theModifiers;
-	int theTrigger;
+    NSInteger theType;
+    NSInteger theCorner;
+    NSInteger theModifiers;
+	NSInteger theTrigger;
     BOOL hoverTriggerDelayed;
     NSImage *myIcon;
     NSString *myLabel;
@@ -23,26 +23,26 @@
 /*-(id)initWithType: (int) type andModifiers: (int) modifiers andString: (NSString *)theString forCorner: (int) corner andClicker:(Clicker *) clicker;
 -(id)initWithType: (int) type andModifiers: (int) modifiers andString: (NSString *)theString forCorner: (int) corner withLabel:(NSString *) label andClicker:(Clicker *) clicker;
 */
--(id)initWithType: (int) type andModifiers: (int) modifiers andTrigger: (int) trigger  isDelayed: (BOOL) hoverTriggerDelayed  andString: (NSString *)theString
-		forCorner: (int)corner withLabel: (NSString *)label andClicker:(Clicker *) clicker;
+-(id)initWithType: (NSInteger) type andModifiers: (NSInteger) modifiers andTrigger: (NSInteger) trigger  isDelayed: (BOOL) hoverTriggerDelayed  andString: (NSString *)theString
+		forCorner: (NSInteger)corner withLabel: (NSString *)label andClicker:(Clicker *) clicker;
 -(void)hideCurrentAction;
 -(void)hideOthersAction;
--(int)type;
--(int)corner;
--(int)modifiers;
+-(NSInteger)type;
+-(NSInteger)corner;
+-(NSInteger)modifiers;
 -(NSString *)string;
 -(NSString *)label;
 -(NSImage *)icon;
 - (NSString *) labelSetting;
--(int)trigger;
+-(NSInteger)trigger;
 - (void) setLabelSetting:(NSString *) label;
 -(void) setString: (NSString *) string;
 -(void) setLabel: (NSString *) label;
 -(void) setIcon: (NSImage *) icon;
--(void) setCorner: (int) corner;
--(void) setType: (int) type;
--(void) setModifiers: (int) modifiers;
--(void) setTrigger: (int) trigger;
+-(void) setCorner: (NSInteger) corner;
+-(void) setType: (NSInteger) type;
+-(void) setModifiers: (NSInteger) modifiers;
+-(void) setTrigger: (NSInteger) trigger;
 -(void) setHoverTriggerDelayed: (BOOL) delayed;
 -(BOOL) hoverTriggerDelayed;
 - (NSComparisonResult)triggerCompare:(ClickAction *)anAction;
@@ -54,9 +54,9 @@
 
 //static
 + (void) logAppleScriptError:(NSDictionary *) err atStep:(NSString *)step;
-+ (NSString *) stringNameForActionType: (int) type;
-+ (NSString *) labelNameForActionType: (int) type;
-+ (BOOL) validActionType: (int) type andString: (NSString *) action;
++ (NSString *) stringNameForActionType: (NSInteger) type;
++ (NSString *) labelNameForActionType: (NSInteger) type;
++ (BOOL) validActionType: (NSInteger) type andString: (NSString *) action;
 
 + (void) exposeAllWindowsAction;
 + (void) exposeApplicationWindowsAction;
