+++ clickBG/ClickWindow.m
@@ -4,7 +4,7 @@
 
 
 // We override this initializer so we can set the NSBorderlessWindowMask styleMask, and set a few other important settings
-- (id)initWithContentRect:(NSRect)contentRect styleMask:(unsigned int)aStyle backing:(NSBackingStoreType)bufferingType defer:(BOOL)flag corner: (int) myCorner 
+- (id)initWithContentRect:(NSRect)contentRect styleMask:(NSUInteger)aStyle backing:(NSBackingStoreType)bufferingType defer:(BOOL)flag corner: (NSInteger) myCorner 
 {
     id win=[super initWithContentRect:contentRect styleMask:aStyle backing:bufferingType defer:flag];    
     if(win){
@@ -33,7 +33,7 @@
     [[NSApp delegate] flagsChanged:theEvent];
 }
 
-- (int) corner
+- (NSInteger) corner
 {
     return corner;
 }
