+++ clickBG/GrayView.h
@@ -19,12 +19,12 @@
     NSDictionary *shadowAttrs;
 	NSArray *actions;
 	NSRect prefFrame;
-    float fadeAlpha;
-    float roundingSize;
-    float insetSize;
-    int pointCorner;
+    CGFloat fadeAlpha;
+    CGFloat roundingSize;
+    CGFloat insetSize;
+    NSInteger pointCorner;
     BOOL dirty;
-    int tailLen;
+    NSInteger tailLen;
     BOOL drawHilite;
 	BOOL showModifierTitle;
 }
@@ -32,24 +32,24 @@
 - (id) initWithFrame: (NSRect) frame andString: (NSString *)msg andImage: (NSImage *) img;
 - (id) initWithFrame: (NSRect) frame andString: (NSString *)msg;
 - (id) initWithFrame: (NSRect) frame andString: (NSString *)msg andImage: (NSImage *) img
-            fadeFrom: (NSColor *)fromCol fadeTo: (NSColor *) toCol cornerSize: (float) cornerSize
-         pointCorner: (int) pCorner;
+            fadeFrom: (NSColor *)fromCol fadeTo: (NSColor *) toCol cornerSize: (CGFloat) cornerSize
+         pointCorner: (NSInteger) pCorner;
 - (void) setFadeFromColor: (NSColor *)color;
 - (void) setFadeToColor: (NSColor *) color;
-- (void) setPointCorner: (int) pCorner;
+- (void) setPointCorner: (NSInteger) pCorner;
 - (void) setActions: (NSArray *) actions;
 - (void) drawFadeFrame: (NSRect)rect;
 - (void) drawGradient: (NSRect) therect fromColor:(NSColor *) from toColor:(NSColor *) to
-            direction: (int) dir;
+            direction: (NSInteger) dir;
 - (void) setDrawString: (NSString *) drawString;
 - (void) setIcon: (NSImage *) icon;
 - (void) setDrawFont:(NSFont *) font color:(NSColor *) color;
 - (void) setDrawHilite:(BOOL)draw;
 - (BOOL) drawHilite;
-- (void) setInsetSize:(float) size;
+- (void) setInsetSize:(CGFloat) size;
 - (void) setShowModifiersTitle: (BOOL) showTitle;
 - (BOOL) showModifiersTitle;
-- (float) insetSize;
+- (CGFloat) insetSize;
 - (void) recalcSize;
 - (NSRect) preferredFrame;
 - (NSRect) calcPreferredFrame;
